+++ Framework/Statistics.php
@@ -217,7 +217,6 @@ class Statistics {
             $sm = new BucketStatistics();
             $sm->add($space, $bucket->getId(), $token);
             Configuration::getLogger()->debug('[stats] bucket created', ['bucket' => $bucket->getId(), 'token' => $token]);
-            // $client->close();
         } catch(Throwable $e) {
             Configuration::getLogger()->error('[stats] createdb error', ['message' => $e->getMessage(), 'line' => $e->getLine(), "file" => $e->getFile(),  'stack' => $e->getTraceAsString()]);
         }