+++ Framework/TableView.php
@@ -207,8 +207,9 @@ class TableView {
      * Generate a basic table view
      * @param array $data table data ( 'key' => value)
      * @param array $headers table headers ( 'key' => 'headername' )
+     * @param bool $small display bootstrap small table
      */
-    public function view($data, $headers) {
+    public function view($data, $headers, $small=false) {
 
         $html = "";
         if ($this->isPrint()) {
@@ -216,27 +217,28 @@ class TableView {
             $html .= "<head>";
             $html .= "<meta charset=\"UTF-8\" />";
             $html .= "<base href=\"" . $rootWeb . "\">";
-            $html .= "<link rel=\"stylesheet\" href=\"/externals/bootstrap/css/bootstrap.min.css\">";
+            $html .= "<link rel=\"stylesheet\" href=\"/externals/node_modules/bootstrap/dist/css/bootstrap.min.css\">";
             $html .= "</head>";
+
         }
         else{
             $html .= $this->addHeader();
         }
 
         if ($this->printAction != "" && $this->exportAction != "" && !$this->isprint) {
-            $html .= "<div class=\"col-xs-2 col-xs-offset-10\">";
-            $html .= "<button type='button' onclick=\"location.href='" . $this->printAction . "?print=1'\" class=\"btn btn-default\">Print</button>";
-            $html .= "<button type='button' onclick=\"location.href='" . $this->exportAction . "?csv=1'\" class=\"btn btn-default\">Export</button>";
+            $html .= "<div class=\"col-2 offset-10\">";
+            $html .= "<button type='button' onclick=\"location.href='" . $this->printAction . "?print=1'\" class=\"btn btn-outline-dark\">Print</button>";
+            $html .= "<button type='button' onclick=\"location.href='" . $this->exportAction . "?csv=1'\" class=\"btn btn-outline-dark\">Export</button>";
             $html .= "</div>";
         } else {
             if ($this->printAction != "" && !$this->isprint) {
-                $html .= "<div class=\"col-xs-2 col-xs-offset-10\">";
-                $html .= "<button type='button' onclick=\"location.href='" . $this->printAction . "?print=1'\" class=\"btn btn-default\">Print</button>";
+                $html .= "<div class=\"col-2 offset-10\">";
+                $html .= "<button type='button' onclick=\"location.href='" . $this->printAction . "?print=1'\" class=\"btn btn-outline-dark\">Print</button>";
                 $html .= "</div>";
             }
             if ($this->exportAction != "" && !$this->isprint) {
-                $html .= "<div class=\"col-xs-2 col-xs-offset-10\">";
-                $html .= "<button type='button' onclick=\"location.href='" . $this->exportAction . "?csv=1'\" class=\"btn btn-default\">Export</button>";
+                $html .= "<div class=\"col-2 offset-10\">";
+                $html .= "<button type='button' onclick=\"location.href='" . $this->exportAction . "?csv=1'\" class=\"btn btn-outline-dark\">Export</button>";
                 $html .= "</div>";
             }
         }
@@ -247,8 +249,11 @@ class TableView {
             $html .= "</div>";
         }
 
-       
-        $html .= "<table id=\"".$this->tableID."\" class=\"table table-bordered table-striped\" cellspacing=\"0\" width=\"100%\">";
+        $is_small = '';
+       if($small) {
+            $is_small = ' table-sm ';
+       }
+        $html .= "<div class=\"table-responsive\"><table id=\"".$this->tableID."\" class=\"table $is_small table-bordered table-striped\" cellspacing=\"0\" width=\"100%\">";
 
         $isButtons = $this->isButtons();
         // table header
@@ -287,7 +292,7 @@ class TableView {
                 if (count($this->linesButtonActions) > 0 && !$this->isprint) {
                     for ($lb = 0; $lb < count($this->linesButtonActions); $lb++) {
                         $html .= '<td style="width: 1%; white-space: nowrap;">';
-                        $html .= "<button type='button' onclick=\"location.href='" . $this->linesButtonActions[$lb] . "/" . $dat[$this->linesButtonActionsIndex[$lb]] . "'\" class=\"btn btn-xs btn-default\">" . $this->linesButtonName[$lb] . "</button><span> </span>";
+                        $html .= "<button type='button' onclick=\"location.href='" . $this->linesButtonActions[$lb] . "/" . $dat[$this->linesButtonActionsIndex[$lb]] . "'\" class=\"btn btn-sm btn-outline-dark\">" . $this->linesButtonName[$lb] . "</button><span> </span>";
                         $html .= "</td>";
                     }
                 }
@@ -299,10 +304,14 @@ class TableView {
                     }
                     $html .= '<td style="width: 1%; white-space: nowrap;">';
                     if($this->editJS){
-                        $html .= "<button onclick=\"editentry('".$this->editURL . "_" . $idxVal."')\" id=\"".$this->editURL . "_" . $idxVal."\" type='button' class=\"btn btn-xs btn-primary\">Edit</button><span> </span>" ;
+                        $html .= "<button onclick=\"editentry('".$this->editURL . "_" . $idxVal."')\" id=\"".$this->editURL . "_" . $idxVal."\" type='button' class=\"btn btn-sm btn-primary\">Edit</button><span> </span>" ;
                     }
                     else{
-                         $html .= "<button type='button' onclick=\"location.href='" . $this->editURL . "/" . $idxVal . "'\" class=\"btn btn-xs btn-primary\">Edit</button><span> </span>";   
+                        $edit_url = $this->editURL . "/" . $idxVal;
+                        if($this->editIndex != "" && str_contains($this->editURL, '{{'.$this->editIndex.'}}')) {
+                            $edit_url = str_replace('{{'.$this->editIndex.'}}', $idxVal, $this->editURL);
+                        }
+                        $html .= "<button type='button' onclick=\"location.href='" . $edit_url . "'\" class=\"btn btn-sm btn-primary\">Edit</button><span> </span>";
                     }  
                     $html .= "</td>";
                 }
@@ -318,7 +327,7 @@ class TableView {
                 }
                 foreach ($headers as $key => $value) {
 
-                    $ccolor = Constants::COLOR_WHITE;
+                    $ccolor = '';
                     if (isset($this->colorIndexes[$key])){  
                         $ccolor = $dat[$this->colorIndexes[$key]];
                     }
@@ -327,7 +336,7 @@ class TableView {
                             $ccolor = $dat[$this->colorIndexes["all"]];
 	    		        }
                     }
-                    $tcolor = Constants::COLOR_BLACK;
+                    $tcolor = '';
                     if(isset($this->colorIndexes["all_text"])) {
                         $tcolor = $dat[$this->colorIndexes["all_text"]];
                     }
@@ -349,7 +358,7 @@ class TableView {
                         else if ($value["type"] == "download"){
                             $html .= '<td>';
                             if ( $val != "" ){
-                                $html .= sprintf('<a target="_blank" rel="noreferrer,noopener"  href="%s"><button type="btn btn-default">%s</button></a>', $val, $value["text"]);
+                                $html .= sprintf('<a target="_blank" rel="noreferrer,noopener"  href="%s"><button type="btn btn-outline-dark">%s</button></a>', $val, $value["text"]);
                             }
                             $html .= '</td>';
                         }
@@ -365,7 +374,7 @@ class TableView {
             }
         }
         $html .= "</tbody>";
-        $html .= "</table>";
+        $html .= "</table></div>";
         if($addDelete) {
             $html .= $this->addDeleteScript();
         }
@@ -424,7 +433,7 @@ class TableView {
         if(!$url) {
             return '<td></td>';
         }
-        $html = "<td>" . "<button type='button' onclick=\"location.href='" . $url . "'\" class=\"btn btn-xs btn-default\"> <span class=\"glyphicon glyphicon-open\" aria-hidden=\"true\"></span> </button>" . "</td>";
+        $html = "<td>" . "<button type='button' onclick=\"location.href='" . $url . "'\" class=\"btn btn-sm btn-outilne-dark\"> <span class=\"bi-download\" aria-hidden=\"true\"></span> </button>" . "</td>";
 
         return $html;
     }
@@ -436,7 +445,7 @@ class TableView {
      * @return string
      */
     private function addDeleteButtonHtml($id, $name) {
-        return "<input class=\"btn btn-xs btn-danger\" type=\"button\" onclick=\"ConfirmDelete($id, '$name')\" value=\"Delete\">";
+        return "<input class=\"btn btn-sm btn-danger\" type=\"button\" onclick=\"ConfirmDelete($id, '$name')\" value=\"Delete\">";
     }
 
     /**