+++ Modules/booking/View/Booking/bookmonth.php
@@ -12,10 +12,12 @@ require_once 'Modules/booking/View/Booking/agendafunction.php';
 <link href="Modules/booking/Theme/styleagenda.css" rel="stylesheet" type="text/css" />
 
 <div class="row" style="background-color: #ffffff;">
-    <div class="col-xs-12">
+    <div class="col-12">
     <?php
 
-    $from = ["month", $date, $bk_id_resource, $bk_id_area, $id_user];
+    $from = ["month", $date, $bk_id_resource, $bk_id_area, $id_user, $detailedView ? 'detailed' : 'simple'];
+    if($bk_id_area == null) { $bk_id_area = '';}
+    if($bk_id_resource == null) { $bk_id_resource = '';}
 
     $nav = [
         'date' => $date,
@@ -25,7 +27,7 @@ require_once 'Modules/booking/View/Booking/agendafunction.php';
         'bk_id_resource' => $bk_id_resource,
         'id_user' => $id_user
     ];
-    echo drawNavigation('month', $id_space, $date, null, $beforeDate, $afterDate, $bk_id_resource, $bk_id_area, $id_user, $lang);
+    echo drawNavigation('month', $id_space, $date, null, $beforeDate, $afterDate, $bk_id_resource, $bk_id_area, $id_user, $detailedView, $lang);
 
     drawAgenda($id_space, $lang, $month, $year, $calEntries, $resourcesBase, $agendaStyle, $resourceInfo, $nav, $from, $context['role']);
     ?>
@@ -33,7 +35,7 @@ require_once 'Modules/booking/View/Booking/agendafunction.php';
 </div>
 
 <div class="row">
-    <div class="col-xs-12" style="background-color: #ffffff;">
+    <div class="col-12" style="background-color: #ffffff;">
     <?php include "Modules/booking/View/colorcodenavbar.php"; ?>
     </div>
 </div>