+++ Modules/core/View/Corespace/configusersAction.php
@@ -3,10 +3,10 @@
     
 <?php startblock('content') ?>
 <div class="row">
-<div class="col-md-12 pm-form-short">
+<div class="col-12 pm-form-short">
     <?php echo $userForm ?>
 </div>
-<div class="col-md-12 pm-table-short">
+<div class="col-12 pm-table-short">
     <?php echo $userTable ?>
 </div>
 </div>