+++ Modules/services/ServicesRouting.php
@@ -80,6 +80,7 @@ class ServicesRouting extends Routing {
 
 
         // stats
+        // deprecated ?
         $this->addRoute("servicesbalance", "servicesbalance", "servicesbalance", "index", array("id_space"), array(""));
 
         // invoicing